% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Extinction.R
\name{extinction}
\alias{extinction}
\title{Performs an extinction sequence}
\usage{
extinction(x, comp = F, nsims = 500, extprob = 1/nrow(x))
}
\arguments{
\item{comp}{compensation yes: comp = T; no: comp = F}

\item{nsims}{of simulation, nsims: numeric}

\item{extprob}{probability of extinction: extprob}
}
\value{
List with how much biomass will be lost with the doomed species and Increase all biomass values by same proportion to maintain initial community biomass
}
\description{
performs an extinction sequence according to a predefined probability order
}
\examples{
extinction(x, comp = F, nsims = 500, extprob = 1/nrow(x)

}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Clement Garcia, \email{broman@wisc.edu}
}
\keyword{hplot}
